requirements.txt
fastapi
uvicorn
qdrant-client
langchain
sentence-transformers
pdfminer.six
python-docx
unstructured
